+++ src/Pim/Bundle/EnrichBundle/Controller/GroupTypeController.php
@@ -4,6 +4,7 @@ namespace Pim\Bundle\EnrichBundle\Controller;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
 use Oro\Bundle\SecurityBundle\Annotation\AclAncestor;
+use Pim\Component\Resource\Model\RemoverInterface;
 use Pim\Bundle\CatalogBundle\Entity\GroupType;
 use Pim\Bundle\EnrichBundle\AbstractController\AbstractDoctrineController;
 use Pim\Bundle\EnrichBundle\Exception\DeleteException;
@@ -35,6 +36,9 @@ class GroupTypeController extends AbstractDoctrineController
     /** @var Form */
     protected $groupTypeForm;
 
+    /** @var RemoverInterface */
+    protected $groupTypeRemover;
+
     /**
      * Constructor
      *
@@ -49,6 +53,7 @@ class GroupTypeController extends AbstractDoctrineController
      * @param ManagerRegistry          $doctrine
      * @param HandlerInterface         $groupTypeHandler
      * @param Form                     $groupTypeForm
+     * @param RemoverInterface         $groupTypeRemover
      */
     public function __construct(
         Request $request,
@@ -61,7 +66,8 @@ class GroupTypeController extends AbstractDoctrineController
         EventDispatcherInterface $eventDispatcher,
         ManagerRegistry $doctrine,
         HandlerInterface $groupTypeHandler,
-        Form $groupTypeForm
+        Form $groupTypeForm,
+        RemoverInterface $groupTypeRemover
     ) {
         parent::__construct(
             $request,
@@ -77,6 +83,7 @@ class GroupTypeController extends AbstractDoctrineController
 
         $this->groupTypeHandler = $groupTypeHandler;
         $this->groupTypeForm    = $groupTypeForm;
+        $this->groupTypeRemover = $groupTypeRemover;
     }
 
     /**
@@ -160,7 +167,7 @@ class GroupTypeController extends AbstractDoctrineController
         } elseif (count($groupType->getGroups()) > 0) {
             throw new DeleteException($this->getTranslator()->trans('flash.group type.cant remove used'));
         } else {
-            $this->remove($groupType);
+            $this->groupTypeRemover->remove($groupType);
         }
 
         if ($this->getRequest()->isXmlHttpRequest()) {